+++ src/grammar/contemplate.js
@@ -68,7 +68,7 @@ var contemplate_grammar = {
                                 "extends", "block", "endblock", "super", "getblock", "include",
                                 "if", "elseif", "elif", "else", "endif", "fi", 
                                 "for", "elsefor", "endfor", "as", "in",
-                                "set", "unset", "isset", "empty", "iif", "continue", "break"
+                                "set", "unset", "isset", "empty", "iif", "continue", "break", "local_set"
                                 ]}
     ,"function"                 : {"autocomplete":true,"tokens":[
                                 "n","s","f","q","qq","dq","addslashes","stripslashes","sprintf",
@@ -90,7 +90,7 @@ var contemplate_grammar_compat = {
                                 "%extends", "%block", "%endblock", "%super", "%getblock", "%include",
                                 "%if", "%elseif", "%elif", "%else", "%endif", "%fi", 
                                 "%for", "%elsefor", "%endfor", "as", "in",
-                                "%set", "%unset", "%isset", "%empty", "%iif", "%continue", "%break"
+                                "%set", "%unset", "%isset", "%empty", "%iif", "%continue", "%break", "%local_set"
                                 ]}
     ,"function"                 : {"autocomplete":true,"tokens":[
                                 "%n","%s","%f","%q","%qq","%dq","%addslashes","%stripslashes","%sprintf",