+++ tests/_tplcache/demo_tpl__global.js
@@ -232,16 +232,34 @@ function Contemplate_demo__global( id )
         __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>Inline (ternary) IF</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% iif( 1+1 == 2, "1+1 = 2", "1+1 = 1" ) %&gt;' + "\n" + '    </pre><br />' + "\n" + '    ' + (((1+1 == 2)?("1+1 = 2"):("1+1 = 1"))) + '' + "\n" + '    ' + "\n" + '    <pre>' + "\n" + '    &lt;% iif( 1+1 == 1, "1+1 = 1", "1+1 = 2" ) %&gt;' + "\n" + '    </pre><br />' + "\n" + '    ' + (((1+1 == 1)?("1+1 = 1"):("1+1 = 2"))) + '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>Inline (ternary) IF (2)</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% iif( !empty($undefined_variable), $undefined_variable, "test with undefined variable passed" ) %&gt;' + "\n" + '    </pre><br />' + "\n" + '    ' + (((!(("undefined" === typeof(data.undefined_variable)) || (null === data.undefined_variable) || Contemplate.empty(data.undefined_variable)))?(data.undefined_variable):("test with undefined variable passed"))) + '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>SET a new tpl variable</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% set($foo, "123") %&gt;' + "\n" + '    </pre><br />' + "\n" + '    ';
         data.foo = ("123");
         
+        __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>SET a new (local) tpl variable</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% local_set($foo_loc, 456) %&gt;' + "\n" + '    &lt;% set($foo_loc, $foo_loc+1) %&gt;' + "\n" + '    </pre><br />' + "\n" + '    ';
+        var _loc_foo_loc = (456);
+        
+        __p__ += '' + "\n" + '    ';
+        _loc_foo_loc = (_loc_foo_loc+1);
+        
         __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>CHECK ISSET for a tpl variable</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% if( isset($foo) ) %&gt;' + "\n" + '        $foo is SET' + "\n" + '    &lt;% else %&gt;' + "\n" + '        $foo is NOT SET' + "\n" + '    &lt;% fi %&gt;' + "\n" + '    </pre><br />' + "\n" + '    ';
         if (("undefined" !== typeof(data.foo) && null !== data.foo))
         {
             
-            __p__ += '' + "\n" + '        $foo is SET' + "\n" + '    ';
+            __p__ += '' + "\n" + '        $foo = ' + (data.foo) + ', is SET' + "\n" + '    ';
         }
         else
         {
             
             __p__ += '' + "\n" + '        $foo is NOT SET' + "\n" + '    ';
+        }
+        
+        __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '' + "\n" + '    <strong>CHECK ISSET for a (local) tpl variable</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% if( isset($foo_loc) ) %&gt;' + "\n" + '        $foo_loc is SET' + "\n" + '    &lt;% else %&gt;' + "\n" + '        $foo_loc is NOT SET' + "\n" + '    &lt;% fi %&gt;' + "\n" + '    </pre><br />' + "\n" + '    ';
+        if (("undefined" !== typeof(_loc_foo_loc) && null !== _loc_foo_loc))
+        {
+            
+            __p__ += '' + "\n" + '        $foo_loc = ' + (_loc_foo_loc) + ', is SET' + "\n" + '    ';
+        }
+        else
+        {
+            
+            __p__ += '' + "\n" + '        $foo_loc is NOT SET' + "\n" + '    ';
         }
         
         __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '' + "\n" + '    <strong>CHECK EMPTY for a tpl variable</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% if( empty($foo) ) %&gt;' + "\n" + '        $foo is EMPTY' + "\n" + '    &lt;% else %&gt;' + "\n" + '        $foo is NOT EMPTY' + "\n" + '    &lt;% fi %&gt;' + "\n" + '    </pre><br />' + "\n" + '    ';
@@ -253,19 +271,31 @@ function Contemplate_demo__global( id )
         else
         {
             
-            __p__ += '' + "\n" + '        $foo is NOT EMPTY' + "\n" + '    ';
+            __p__ += '' + "\n" + '        $foo = ' + (data.foo) + ', is NOT EMPTY' + "\n" + '    ';
+        }
+        
+        __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '' + "\n" + '    <strong>CHECK EMPTY for a (local) tpl variable</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% if( empty($foo_loc) ) %&gt;' + "\n" + '        $foo_loc is EMPTY' + "\n" + '    &lt;% else %&gt;' + "\n" + '        $foo_loc is NOT EMPTY' + "\n" + '    &lt;% fi %&gt;' + "\n" + '    </pre><br />' + "\n" + '    ';
+        if ((("undefined" === typeof(_loc_foo_loc)) || (null === _loc_foo_loc) || Contemplate.empty(_loc_foo_loc)))
+        {
+            
+            __p__ += '' + "\n" + '        $foo_loc is EMPTY' + "\n" + '    ';
+        }
+        else
+        {
+            
+            __p__ += '' + "\n" + '        $foo_loc = ' + (_loc_foo_loc) + ', is NOT EMPTY' + "\n" + '    ';
         }
         
         __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '' + "\n" + '    <!-- include a (sub-)template file -->' + "\n" + '    <strong>INCLUDE a (sub-)template file</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% include("date") %&gt;' + "\n" + '    </pre><br />' + "\n" + '     <!-- print a localized date php-style -->' + "\n" + '<strong>A (localized) date, PHP-style</strong><br />' + "\n" + '' + (Contemplate.ldate("M, D, d")) + '' + "\n" + '';
         
         __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '' + "\n" + '    <strong>CALL another (sub-)template</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% for($users as $i=>$usergroup) %&gt;' + "\n" + '        &lt;!-- call a (sub-)template --&gt;' + "\n" + '        &lt;% tpl("sub", {"i" : $i, "users" : $users}) %&gt;' + "\n" + '    &lt;% endfor %&gt;' + "\n" + '    </pre><br />' + "\n" + '    ';
-        var _loc_147 = data.users, _loc_148 = _loc_147 ? Object.keys(_loc_147) : null,
-            _loc_149, _loc_i, _loc_usergroup, _loc_150 = _loc_148 ? _loc_148.length : 0;
-        if (_loc_150)
+        var _loc_156 = data.users, _loc_157 = _loc_156 ? Object.keys(_loc_156) : null,
+            _loc_158, _loc_i, _loc_usergroup, _loc_159 = _loc_157 ? _loc_157.length : 0;
+        if (_loc_159)
         {
-            for (_loc_149=0; _loc_149<_loc_150; _loc_149++)
+            for (_loc_158=0; _loc_158<_loc_159; _loc_158++)
             {
-                _loc_i = _loc_148[_loc_149]; _loc_usergroup = _loc_147[_loc_i];
+                _loc_i = _loc_157[_loc_158]; _loc_usergroup = _loc_156[_loc_i];
                 
                 
                 __p__ += '' + "\n" + '        <!-- call a (sub-)template -->' + "\n" + '        ' + (Contemplate.tpl("sub", {"i" : _loc_i, "users" : data.users})) + '' + "\n" + '    ';