+++ tests/_tplcache/demo_tpl__global.py
@@ -140,13 +140,27 @@ def __getTplClass__(Contemplate):
             __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>Inline (ternary) IF</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% iif( 1+1 == 2, "1+1 = 2", "1+1 = 1" ) %&gt;' + "\n" + '    </pre><br />' + "\n" + '    ' + str((("1+1 = 2") if (1+1 == 2) else ("1+1 = 1"))) + '' + "\n" + '    ' + "\n" + '    <pre>' + "\n" + '    &lt;% iif( 1+1 == 1, "1+1 = 1", "1+1 = 2" ) %&gt;' + "\n" + '    </pre><br />' + "\n" + '    ' + str((("1+1 = 1") if (1+1 == 1) else ("1+1 = 2"))) + '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>Inline (ternary) IF (2)</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% iif( !empty($undefined_variable), $undefined_variable, "test with undefined variable passed" ) %&gt;' + "\n" + '    </pre><br />' + "\n" + '    ' + str(((data['undefined_variable']) if (not (("undefined_variable" not in data) or (data['undefined_variable'] is None) or Contemplate.empty(data['undefined_variable']))) else ("test with undefined variable passed"))) + '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>SET a new tpl variable</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% set($foo, "123") %&gt;' + "\n" + '    </pre><br />' + "\n" + '    '
             data['foo'] = ("123")
              
+            __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>SET a new (local) tpl variable</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% local_set($foo_loc, 456) %&gt;' + "\n" + '    &lt;% set($foo_loc, $foo_loc+1) %&gt;' + "\n" + '    </pre><br />' + "\n" + '    '
+            _loc_foo_loc = (456)
+             
+            __p__ += '' + "\n" + '    '
+            _loc_foo_loc = (_loc_foo_loc+1)
+             
             __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>CHECK ISSET for a tpl variable</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% if( isset($foo) ) %&gt;' + "\n" + '        $foo is SET' + "\n" + '    &lt;% else %&gt;' + "\n" + '        $foo is NOT SET' + "\n" + '    &lt;% fi %&gt;' + "\n" + '    </pre><br />' + "\n" + '    '
             if ((("foo" in data) and (data['foo'] is not None))):
                  
-                __p__ += '' + "\n" + '        $foo is SET' + "\n" + '    '
+                __p__ += '' + "\n" + '        $foo = ' + str(data['foo']) + ', is SET' + "\n" + '    '
             else:
                  
                 __p__ += '' + "\n" + '        $foo is NOT SET' + "\n" + '    '
+             
+            __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '' + "\n" + '    <strong>CHECK ISSET for a (local) tpl variable</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% if( isset($foo_loc) ) %&gt;' + "\n" + '        $foo_loc is SET' + "\n" + '    &lt;% else %&gt;' + "\n" + '        $foo_loc is NOT SET' + "\n" + '    &lt;% fi %&gt;' + "\n" + '    </pre><br />' + "\n" + '    '
+            if ((("_loc_foo_loc" in locals()) and (_loc_foo_loc is not None))):
+                 
+                __p__ += '' + "\n" + '        $foo_loc = ' + str(_loc_foo_loc) + ', is SET' + "\n" + '    '
+            else:
+                 
+                __p__ += '' + "\n" + '        $foo_loc is NOT SET' + "\n" + '    '
              
             __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '' + "\n" + '    <strong>CHECK EMPTY for a tpl variable</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% if( empty($foo) ) %&gt;' + "\n" + '        $foo is EMPTY' + "\n" + '    &lt;% else %&gt;' + "\n" + '        $foo is NOT EMPTY' + "\n" + '    &lt;% fi %&gt;' + "\n" + '    </pre><br />' + "\n" + '    '
             if ((("foo" not in data) or (data['foo'] is None) or Contemplate.empty(data['foo']))):
@@ -154,7 +168,15 @@ def __getTplClass__(Contemplate):
                 __p__ += '' + "\n" + '        $foo is EMPTY' + "\n" + '    '
             else:
                  
-                __p__ += '' + "\n" + '        $foo is NOT EMPTY' + "\n" + '    '
+                __p__ += '' + "\n" + '        $foo = ' + str(data['foo']) + ', is NOT EMPTY' + "\n" + '    '
+             
+            __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '' + "\n" + '    <strong>CHECK EMPTY for a (local) tpl variable</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% if( empty($foo_loc) ) %&gt;' + "\n" + '        $foo_loc is EMPTY' + "\n" + '    &lt;% else %&gt;' + "\n" + '        $foo_loc is NOT EMPTY' + "\n" + '    &lt;% fi %&gt;' + "\n" + '    </pre><br />' + "\n" + '    '
+            if ((("_loc_foo_loc" not in locals()) or (_loc_foo_loc is None) or Contemplate.empty(_loc_foo_loc))):
+                 
+                __p__ += '' + "\n" + '        $foo_loc is EMPTY' + "\n" + '    '
+            else:
+                 
+                __p__ += '' + "\n" + '        $foo_loc = ' + str(_loc_foo_loc) + ', is NOT EMPTY' + "\n" + '    '
              
             __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '' + "\n" + '    <!-- include a (sub-)template file -->' + "\n" + '    <strong>INCLUDE a (sub-)template file</strong><br />' + "\n" + '    <pre>' + "\n" + '    &lt;% include("date") %&gt;' + "\n" + '    </pre><br />' + "\n" + '     <!-- print a localized date php-style -->' + "\n" + '<strong>A (localized) date, PHP-style</strong><br />' + "\n" + '' + str(Contemplate.ldate("M, D, d")) + '' + "\n" + ''