+++ changelog.md
@@ -3,6 +3,7 @@
 __1.1.4__
 * lazier init for faster performance
 * new function `is_array` to test both if is `array` (`strict`) or is `object`
+* new directive `local_set` to set/update a **local** variable
 
 
 __1.1.3__